Re-write local state query client code with LSQ monad
Goals
- Make AcquireFailures impossible (for wallet API users)
- Eliminate integration test flakiness
- Reduce boilerplate

by adding a monad for composing queries to be run against any, but the
same, node tip.

* wip: leverage STM retrying

This adopts the state observation pattern Duncan described in a recent
STM seminar.

We now use pure TVar's. Not TMVar or Chan. We can still efficiently
observe for changes by using `guard`.

This seems to make the tests much greener. Hoping this fixes several
flaky tests on master.

* Reduce boilerplate, and re-add support for past eras